requirements.txt
aiosmtplib~=2.0.2
asyncio==3.4.3
bcrypt==3.2.2
fastapi~=0.111.0
itsdangerous~=2.2.0
Jinja2==3.1.2
jwt==1.3.1
passlib==1.7.4
pydantic~=2.7.1
redis~=4.6.0
SQLAlchemy~=2.0.30
starlette~=0.37.2
stripe~=10.1.0
uvicorn==0.18.2
aiofiles~=24.1.0
celery~=5.4.0
PyJWT~=2.8.0
alembic~=1.13.1
python-dotenv~=1.0.1